+++ es/utils/aepp-wallet-communication/rpc/aepp-rpc.js
@@ -41,10 +41,10 @@ const RESPONSES = {
     },
   [METHODS.aepp.subscribeAddress]: (instance) =>
     (msg) => {
-      if (
-        msg.result &&
-        Object.prototype.hasOwnProperty.call(msg.result, 'address')
-      ) instance.rpcClient.accounts = msg.result.address
+      if (msg.result) {
+        msg.result.address && (instance.rpcClient.accounts = msg.result.address)
+        msg.result.subscription && (instance.rpcClient.addressSubscription = msg.result.subscription)
+      }
 
       instance.rpcClient.processResponse(msg, ({ id, result }) => [result])
     },
@@ -106,6 +106,10 @@ export const AeppRpc = Ae.compose({
   methods: {
     sign () {
     },
+    addresses () {
+      if (!this.rpcClient.currentAccount) throw new Error('You do not subscribed for account.')
+      return [this.rpcClient.currentAccount, ...Object.keys(this.rpcClient.accounts.connected)]
+    },
     /**
      * Connect to wallet
      * @function connectToWallet
@@ -134,15 +138,17 @@ export const AeppRpc = Ae.compose({
      * @return {void}
      */
     async disconnectWallet (sendDisconnect = true) {
-      if (!this.rpcClient || !this.rpcClient.connection.isConnected() || !this.rpcClient.isConnected()) throw new Error('You are not connected to Wallet')
+      if (!this.rpcClient || !this.rpcClient.isConnected()) throw new Error('You are not connected to Wallet')
       sendDisconnect && this.rpcClient.sendMessage(message(METHODS.closeConnection, { reason: 'bye' }), true)
       this.rpcClient.disconnect()
       this.rpcClient = null
     },
     async address ({ onAccount } = {}) {
-      if (!this.rpcClient || !this.rpcClient.connection.isConnected() || !this.rpcClient.isConnected()) throw new Error('You are not connected to Wallet')
-      if (!this.rpcClient.getCurrentAccount()) throw new Error('You do not subscribed for account.')
-      return this.rpcClient.getCurrentAccount({ onAccount })
+      if (!this.rpcClient || !this.rpcClient.isConnected()) throw new Error('You are not connected to Wallet')
+      const current = this.rpcClient.currentAccount
+      if (!current) throw new Error('You do not subscribed for accounts.')
+      if (onAccount && !this.rpcClient.hasAccessToAccount(onAccount)) throw new Error(`You are not have access to account ${onAccount}`)
+      return onAccount || current
     },
     /**
      * Ask address from wallet
@@ -152,11 +158,9 @@ export const AeppRpc = Ae.compose({
      * @return {Promise} Address from wallet
      */
     async askAddresses () {
-      if (!this.rpcClient || !this.rpcClient.connection.isConnected() || !this.rpcClient.isConnected()) throw new Error('You are not connected to Wallet')
-      if (!this.rpcClient.getCurrentAccount()) throw new Error('You do not subscribed for account.')
-      return this.rpcClient.addCallback(
-        this.rpcClient.sendMessage(message(METHODS.aepp.address))
-      )
+      if (!this.rpcClient || !this.rpcClient.isConnected()) throw new Error('You are not connected to Wallet')
+      if (!this.rpcClient.currentAccount) throw new Error('You do not subscribed for account.')
+      return this.rpcClient.request(METHODS.aepp.address)
     },
     /**
      * Subscribe for addresses from wallet
@@ -168,10 +172,8 @@ export const AeppRpc = Ae.compose({
      * @return {Promise} Address from wallet
      */
     async subscribeAddress (type, value) {
-      if (!this.rpcClient || !this.rpcClient.connection.isConnected() || !this.rpcClient.isConnected()) throw new Error('You are not connected to Wallet')
-      return this.rpcClient.addCallback(
-        this.rpcClient.sendMessage(message(METHODS.aepp.subscribeAddress, { type, value }))
-      )
+      if (!this.rpcClient || !this.rpcClient.isConnected()) throw new Error('You are not connected to Wallet')
+      return this.rpcClient.request(METHODS.aepp.subscribeAddress, { type, value })
     },
     /**
      * Overwriting of `signTransaction` AE method
@@ -182,11 +184,10 @@ export const AeppRpc = Ae.compose({
      * @return {Promise<String>} Signed transaction
      */
     async signTransaction (tx, opt = {}) {
-      if (!this.rpcClient || !this.rpcClient.connection.isConnected() || !this.rpcClient.isConnected()) throw new Error('You are not connected to Wallet')
-      if (!this.rpcClient.getCurrentAccount()) throw new Error('You do not subscribed for account.')
-      return this.rpcClient.addCallback(
-        this.rpcClient.sendMessage(message(METHODS.aepp.sign, { ...opt, tx, returnSigned: true }))
-      )
+      if (!this.rpcClient || !this.rpcClient.isConnected()) throw new Error('You are not connected to Wallet')
+      if (!this.rpcClient.currentAccount) throw new Error('You do not subscribed for account.')
+      if (opt.onAccount && !this.rpcClient.hasAccessToAccount(opt.onAccount)) throw new Error(`You are not have access to account ${opt.onAccount}`)
+      return this.rpcClient.request(METHODS.aepp.sign, { ...opt, tx, returnSigned: true })
     },
     /**
      * Overwriting of `signMessage` AE method
@@ -197,11 +198,10 @@ export const AeppRpc = Ae.compose({
      * @return {Promise<String>} Signed transaction
      */
     async signMessage (msg, opt = {}) {
-      if (!this.rpcClient || !this.rpcClient.connection.isConnected() || !this.rpcClient.isConnected()) throw new Error('You are not connected to Wallet')
-      if (!this.rpcClient.getCurrentAccount()) throw new Error('You do not subscribed for account.')
-      return this.rpcClient.addCallback(
-        this.rpcClient.sendMessage(message(METHODS.aepp.signMessage, { ...opt, message: msg }))
-      )
+      if (!this.rpcClient || !this.rpcClient.isConnected()) throw new Error('You are not connected to Wallet')
+      if (!this.rpcClient.currentAccount) throw new Error('You do not subscribed for account.')
+      if (opt.onAccount && !this.rpcClient.hasAccessToAccount(opt.onAccount)) throw new Error(`You are not have access to account ${opt.onAccount}`)
+      return this.rpcClient.request(METHODS.aepp.signMessage, { ...opt, message: msg })
     },
     /**
      * Send connection request to wallet
@@ -211,12 +211,12 @@ export const AeppRpc = Ae.compose({
      * @return {Promise} Connection response
      */
     async sendConnectRequest () {
-      return this.rpcClient.addCallback(
-        this.rpcClient.sendMessage(message(METHODS.aepp.connect, {
+      return this.rpcClient.request(
+        METHODS.aepp.connect, {
           name: this.name,
           version: VERSION,
           networkId: this.getNetworkId()
-        }))
+        }
       )
     },
     /**
@@ -232,16 +232,15 @@ export const AeppRpc = Ae.compose({
      * @return {Promise<Object>} Transaction broadcast result
      */
     async send (tx, options = {}) {
-      if (!this.rpcClient || !this.rpcClient.connection.isConnected() || !this.rpcClient.isConnected()) throw new Error('You are not connected to Wallet')
-      if (!this.rpcClient.getCurrentAccount()) throw new Error('You do not subscribed for account.')
+      if (!this.rpcClient || !this.rpcClient.isConnected()) throw new Error('You are not connected to Wallet')
+      if (!this.rpcClient.currentAccount) throw new Error('You do not subscribed for account.')
+      if (options.onAccount && !this.rpcClient.hasAccessToAccount(options.onAccount)) throw new Error(`You are not have access to account ${options.onAccount}`)
       const opt = R.merge(this.Ae.defaults, options)
       if (!opt.walletBroadcast) {
         const signed = await this.signTransaction(tx, { onAccount: opt.onAccount })
         return this.sendTransaction(signed, opt)
       }
-      return this.rpcClient.addCallback(
-        this.rpcClient.sendMessage(message(METHODS.aepp.sign, { onAccount: opt.onAccount, tx, returnSigned: false }))
-      )
+      return this.rpcClient.request(METHODS.aepp.sign, { onAccount: opt.onAccount, tx, returnSigned: false })
     }
   }
 })