+++ es/utils/aepp-wallet-communication/rpc/wallet-rpc.js
@@ -8,10 +8,11 @@
 import Ae from '../../../ae'
 import Accounts from '../../../accounts'
 import Selector from '../../../account/selector'
+import TxObject from '../../../tx/tx-object'
 
 import { RpcClients } from './rpc-clients'
-import { getBrowserAPI, getHandler, message, sendResponseMessage } from '../helpers'
-import { ERRORS, METHODS, RPC_STATUS, VERSION, WALLET_TYPE, SUBSCRIPTION_VALUES } from '../schema'
+import { getBrowserAPI, getHandler, isValidAccounts, message, resolveOnAccount, sendResponseMessage } from '../helpers'
+import { ERRORS, METHODS, RPC_STATUS, VERSION, WALLET_TYPE } from '../schema'
 import { v4 as uuid } from 'uuid'
 
 const rpcClients = RpcClients()
@@ -29,147 +30,143 @@ const RESPONSES = {}
 const REQUESTS = {
   // Store client info and prepare two fn for each client `connect` and `denyConnection`
   // which automatically prepare and send response for that client
-  [METHODS.aepp.connect]: (instance, { client }) =>
-    ({ id, method, params: { name, networkId, version, icons } }, origin) => {
-      // Check if protocol and network is compatible with wallet
-      if (version !== VERSION) return sendResponseMessage(client)(id, method, { error: ERRORS.unsupportedProtocol() })
-      if (networkId !== instance.getNetworkId()) return sendResponseMessage(client)(id, method, { error: ERRORS.unsupportedNetwork() })
+  async [METHODS.aepp.connect] (callInstance, instance, client, { name, networkId, version, icons }) {
+    // Check if protocol and network is compatible with wallet
+    if (version !== VERSION) return { error: ERRORS.unsupportedProtocol() }
+    if (networkId !== instance.getNetworkId()) return { error: ERRORS.unsupportedNetwork() }
 
-      // Action methods
-      const accept = (id) => () => {
+    // Store new AEPP and wait for connection approve
+    rpcClients.updateClientInfo(client.id, {
+      status: RPC_STATUS.WAITING_FOR_CONNECTION_APPROVE,
+      name,
+      networkId,
+      icons,
+      version
+    })
+
+    // Call onConnection callBack to notice Wallet about new AEPP
+    return callInstance(
+      'onConnection',
+      { name, networkId, version },
+      () => {
         rpcClients.updateClientInfo(client.id, { status: RPC_STATUS.CONNECTED })
-        sendResponseMessage(client)(id, method, { result: instance.getWalletInfo() })
-      }
-      const deny = (id) => (error) => {
+        return { result: instance.getWalletInfo() }
+      },
+      (error) => {
         rpcClients.updateClientInfo(client.id, { status: RPC_STATUS.CONNECTION_REJECTED })
-        sendResponseMessage(client)(id, METHODS.aepp.connect, { error: ERRORS.connectionDeny(error) })
+        return { error: ERRORS.connectionDeny(error) }
       }
+    )
+  },
+  async [METHODS.aepp.subscribeAddress] (callInstance, instance, client, { type, value }) {
+    // Authorization check
+    if (!client.isConnected()) return { error: ERRORS.notAuthorize() }
 
-      // Store new AEPP and wait for connection approve
-      rpcClients.updateClientInfo(client.id, {
-        status: RPC_STATUS.WAITING_FOR_CONNECTION_APPROVE,
-        name,
-        networkId,
-        icons,
-        version
-      })
-
-      // Call onConnection callBack to notice Wallet about new AEPP
-      instance.onConnection(
-        client,
-        client.addAction({
-          id,
-          method,
-          params: { name, networkId, version }
-        }, [accept(id), deny(id)]),
-        origin
-      )
-    },
-  [METHODS.aepp.subscribeAddress]: (instance, { client }) =>
-    ({ id, method, params: { type, value } }, origin) => {
-      // Authorization check
-      if (!client.isConnected()) return sendResponseMessage(client)(id, method, { error: ERRORS.notAuthorize() })
-
-      const accept = (id) =>
-        () => sendResponseMessage(client)(
-          id,
-          method,
-          {
-            result: {
-              subscription: client.updateSubscription(type, value),
-              address: instance.getAccounts()
-            }
-          })
-      const deny = (id) => (error) => sendResponseMessage(client)(id, method, { error: ERRORS.rejectedByUser(error) })
-
-      instance.onSubscription(
-        client,
-        client.addAction({ id, method, params: { type, value } }, [accept(id), deny(id)]),
-        origin
-      )
-    },
-  [METHODS.aepp.address]: (instance, { client }) =>
-    ({ id, method }, origin) => {
-      // Authorization check
-      if (!client.isConnected()) return sendResponseMessage(client)(id, method, { error: ERRORS.notAuthorize() })
-
-      const accept = (id) =>
-        () => sendResponseMessage(client)(
-          id,
-          method,
-          {
-            result: instance.addresses()
-          })
-      const deny = (id) => (error) => sendResponseMessage(client)(id, method, { error: ERRORS.rejectedByUser(error) })
-
-      instance.onAskAccounts(
-        client,
-        client.addAction({ id, method }, [accept(id), deny(id)]),
-        origin
-      )
-    },
-  [METHODS.aepp.sign]: (instance, { client }) =>
-    async ({ id, method, params: { tx, onAccount, returnSigned = false } }, origin) => {
-      // Authorization check
-      if (!client.isConnected()) return sendResponseMessage(client)(id, method, { error: ERRORS.notAuthorize() })
-      // NetworkId check
-      if (client.info.networkId !== instance.getNetworkId()) return sendResponseMessage(client)(id, method, { error: ERRORS.unsupportedNetwork() })
-
-      const accept = (id) => async (rawTx) => {
+    return callInstance(
+      'onSubscription',
+      { type, value },
+      async ({ accounts } = {}) => {
         try {
-          const result = {
+          const clientAccounts = accounts || instance.getAccounts()
+          if (!isValidAccounts(clientAccounts)) throw new Error('Invalid provided accounts object')
+          const subscription = client.updateSubscription(type, value)
+          client.setAccounts(clientAccounts, { forceNotification: true })
+          return {
             result: {
-              ...returnSigned
-                ? { signedTransaction: await instance.signTransaction(rawTx || tx, { onAccount }) }
-                : { transactionHash: await instance.send(rawTx || tx, { onAccount, verify: false }) }
+              subscription,
+              address: clientAccounts
             }
           }
-          sendResponseMessage(client)(
-            id,
-            method,
-            result
-          )
+        } catch (e) {
+          console.error(e)
+          return { error: ERRORS.internalError({ msg: e.message }) }
+        }
+      },
+      (error) => ({ error: ERRORS.rejectedByUser(error) })
+    )
+  },
+  async [METHODS.aepp.address] (callInstance, instance, client) {
+    // Authorization check
+    if (!client.isConnected()) return { error: ERRORS.notAuthorize() }
+    if (!client.isSubscribed()) return { error: ERRORS.notAuthorize() }
+
+    return callInstance(
+      'onAskAccounts',
+      {},
+      ({ accounts } = {}) => ({ result: accounts || [...Object.keys(client.accounts.current), ...Object.keys(client.accounts.connected)] }),
+      (error) => ({ error: ERRORS.rejectedByUser(error) })
+    )
+  },
+  async [METHODS.aepp.sign] (callInstance, instance, client, { tx, onAccount, returnSigned = false }) {
+    const address = onAccount || client.currentAccount
+    // Authorization check
+    if (!client.isConnected()) return { error: ERRORS.notAuthorize() }
+    // Account permission check
+    if (!client.hasAccessToAccount(address)) return { error: ERRORS.permissionDeny({ account: address }) }
+    // NetworkId check
+    if (client.info.networkId !== instance.getNetworkId()) return { error: ERRORS.unsupportedNetwork() }
+
+    return callInstance(
+      'onSign',
+      { tx, returnSigned, onAccount: address, txObject: TxObject.fromString(tx) },
+      async (rawTx, opt = {}) => {
+        let onAcc
+        try {
+          onAcc = resolveOnAccount(instance.addresses(), address, opt)
+        } catch (e) {
+          console.error(e)
+          return { error: ERRORS.internalError({ msg: e.message }) }
+        }
+        try {
+          return {
+            result: {
+              ...returnSigned
+                ? { signedTransaction: await instance.signTransaction(rawTx || tx, { onAccount: onAcc }) }
+                : { transactionHash: await instance.send(rawTx || tx, { onAccount: onAcc, verify: false }) }
+            }
+          }
         } catch (e) {
           if (!returnSigned) {
             // Validate transaction
             const validationResult = await instance.unpackAndVerify(rawTx || tx)
-            if (validationResult.validation.length) return sendResponseMessage(client)(id, method, { error: ERRORS.invalidTransaction(validationResult) })
+            if (validationResult.validation.length) return { error: ERRORS.invalidTransaction(validationResult) }
             // Send broadcast failed error to aepp
-            sendResponseMessage(client)(id, method, { error: ERRORS.broadcastFailde(e.message) })
+            return { error: ERRORS.broadcastFailde(e.message) }
           }
           throw e
         }
-      }
+      },
+      (error) => ({ error: ERRORS.rejectedByUser(error) })
+    )
+  },
+  async [METHODS.aepp.signMessage] (callInstance, instance, client, { message, onAccount }) {
+    // Authorization check
+    if (!client.isConnected()) return { error: ERRORS.notAuthorize() }
+    const address = onAccount || client.currentAccount
+    if (!client.hasAccessToAccount(address)) return { error: ERRORS.permissionDeny({ account: address }) }
 
-      const deny = (id) => (error) => sendResponseMessage(client)(id, method, { error: ERRORS.rejectedByUser(error) })
-
-      instance.onSign(
-        client,
-        client.addAction({ id, method, params: { tx, returnSigned, onAccount } }, [accept(id), deny(id)]),
-        origin
-      )
-    },
-  [METHODS.aepp.signMessage]: (instance, { client }) =>
-    async ({ id, method, params: { message, onAccount } }, origin) => {
-      // Authorization check
-      if (!client.isConnected()) return sendResponseMessage(client)(id, method, { error: ERRORS.notAuthorize() })
-
-      const accept = (id) => async () => sendResponseMessage(client)(
-        id,
-        method,
-        {
-          result: { signature: await instance.signMessage(message, { onAccount, returnHex: true }) }
+    return callInstance(
+      'onMessageSign',
+      { message, onAccount: address },
+      async (opt = {}) => {
+        try {
+          const onAcc = resolveOnAccount(instance.addresses(), address, opt)
+          return {
+            result: {
+              signature: await instance.signMessage(message, {
+                onAccount: onAcc,
+                returnHex: true
+              })
+            }
+          }
+        } catch (e) {
+          console.error(e)
+          return { error: ERRORS.internalError({ msg: e.message }) }
         }
-      )
-
-      const deny = (id) => (error) => sendResponseMessage(client)(id, method, { error: ERRORS.rejectedByUser(error) })
-
-      instance.onMessageSign(
-        client,
-        client.addAction({ id, method, params: { message, onAccount } }, [accept(id), deny(id)]),
-        origin
-      )
-    }
+      },
+      (error) => ({ error: ERRORS.rejectedByUser(error) })
+    )
+  }
 }
 
 const handleMessage = (instance, id) => async (msg, origin) => {
@@ -180,7 +177,18 @@ const handleMessage = (instance, id) => async (msg, origin) => {
   if (Object.prototype.hasOwnProperty.call(client.callbacks, msg.id)) {
     return getHandler(RESPONSES, msg)(instance, { client })(msg, origin)
   } else {
-    return getHandler(REQUESTS, msg)(instance, { client })(msg, origin)
+    const { id, method } = msg
+    const callInstance = (methodName, params, accept, deny) => new Promise(resolve => {
+      instance[methodName](
+        client,
+        client.addAction({ id, method, params }, [
+          (...args) => resolve(accept(...args)), (...args) => resolve(deny(...args))
+        ]),
+        origin
+      )
+    })
+    const response = await getHandler(REQUESTS, msg)(callInstance, instance, client, msg.params)
+    sendResponseMessage(client)(id, method, response)
   }
 }
 
@@ -222,34 +230,48 @@ export const WalletRpc = Ae.compose(Accounts, Selector, {
     const _selectNode = this.selectNode.bind(this)
 
     // Overwrite AE methods
-    this.selectAccount = (address) => {
+    this.selectAccount = (address, { condition = () => true } = {}) => {
       _selectAccount(address)
-      rpcClients.sentNotificationByCondition(
-        message(METHODS.wallet.updateAddress, this.getAccounts()),
-        (client) =>
-          (
-            client.addressSubscription.includes(SUBSCRIPTION_VALUES.current) ||
-            client.addressSubscription.includes(SUBSCRIPTION_VALUES.connected)
-          ) &&
-          client.isConnected())
-    }
-    this.addAccount = async (account, { select, meta } = {}) => {
-      await _addAccount(account, { select })
-      // Send notification 'update.address' to all Aepp which are subscribed for connected accounts
-      rpcClients.sentNotificationByCondition(
-        message(METHODS.wallet.updateAddress, this.getAccounts()),
+      rpcClients.operationByCondition(
         (client) =>
           client.isConnected() &&
-          (
-            client.addressSubscription.includes(SUBSCRIPTION_VALUES.connected) ||
-            (select && client.addressSubscription.includes(SUBSCRIPTION_VALUES.current))
-          )
+          client.isSubscribed() &&
+          client.hasAccessToAccount(address) &&
+          condition(client),
+        (client) =>
+          client.setAccounts({
+            current: { [address]: {} },
+            connected: {
+              ...client.accounts.current,
+              ...Object.entries(client.accounts.connected)
+                .reduce((acc, [k, v]) => ({ ...acc, ...k !== address ? { [k]: v } : {} }), {})
+            }
+          })
+      )
+    }
+    this.addAccount = async (account, { select, meta = {}, condition = () => true } = {}) => {
+      await _addAccount(account, { select })
+      const address = await account.address()
+      // Send notification 'update.address' to all Aepp which are subscribed for connected accounts
+      rpcClients.operationByCondition(
+        (client) =>
+          client.isConnected() &&
+          client.isSubscribed() &&
+          condition(client),
+        (client) =>
+          client.setAccounts({
+            current: { ...select ? { [address]: meta } : client.accounts.current },
+            connected: {
+              ...select ? client.accounts.current : { [address]: meta },
+              ...client.accounts.connected
+            }
+          })
       )
     }
     this.selectNode = (name) => {
       _selectNode(name)
       // Send notification 'update.network' to all Aepp which connected
-      rpcClients.sentNotificationByCondition(
+      rpcClients.sendNotificationByCondition(
         message(METHODS.updateNetwork, { networkId: this.getNetworkId() }),
         (client) => client.isConnected()
       )