+++ es/utils/string.js
@@ -54,7 +54,7 @@ export function pascalToSnake (s) {
  * @return {boolean} True if the string is valid hex, false otherwise.
  */
 export function isHex (str) {
-  return !!(str.length % 2 === 0 && str.match(/^[0-9a-f]+$/i))
+  return !!(typeof str === 'string' && str.length % 2 === 0 && str.match(/^[0-9a-f]+$/i))
 }
 
 /**